refactor: :memo: renames poo verified file name to match component name
